Check .dae file for referenced textures as well
This is equivalent to bmubin's dae fixing step.
Also if the json specifies a MaterialAlb indicating that we'll use texture blending, let's discard whatever's in the json and use what's in the .dae instead. Apparently it's more accurate in such a case, eg. FldObj_MountainLanayru_A-03 (although this object is still a mess because the .dae also insists on Albedo B instead of the correct Albedo A, but what can you do, it's a fancy ice material...)