Fix material de-duplication RGBNode false match
I wasn't adding the default values of iterables to the hash string properly, so when two materials only differed by the color of an RGB node, it was counted as two equal materials.

Since I can't guarantee this doesn't have more mistakes and unaccounted cases, I disabled it by default.